Fix: 사용자 정보 관련 UserInfoService로 분리 + 회원탈퇴 로직 User 도메인에서 수행하도록 수정
