+++ src/main/java/com/toty/user/presentation/UserViewController.java
@@ -1,6 +1,7 @@
 package com.toty.user.presentation;
 
 import com.toty.annotation.CurrentUser;
+import com.toty.user.application.UserInfoService;
 import com.toty.user.application.UserService;
 import com.toty.user.domain.model.User;
 import com.toty.user.dto.response.UserInfoResponse;
@@ -18,24 +19,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class UserViewController {
 
     private final UserService userService;
+    private final UserInfoService userInfoService;
 
     // 정보 수정(View)
     @GetMapping("/edit-form/{id}")
     public String updateProc(@CurrentUser User user, @PathVariable Long id, Model model){
         // 본인인지 확인 -> 아니면 예외
 
-        UserInfoResponse userInfo = userService.getMyInfoForUpdate(user, id);
+        UserInfoResponse userInfo = userInfoService.getMyInfoForUpdate(user, id);
 
         model.addAttribute("userInfo", userInfo);
         return "update";
     }
-
-    // 회원 탈퇴(view)
-    @DeleteMapping("/{id}")
-    public String delete(@CurrentUser User user, @PathVariable Long id) {
-        // 본인 확인 로직 -> 아니면 예외
-        userService.deleteUser(user,id); //soft delete
-        return "redirect:/api/home";
-    }
-
 }
